gRPC/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-auth==2.6.2
googleapis-common-protos==1.56.0
grpc-stubs==1.24.8
grpcio==1.44.0
grpcio-tools==1.44.0
idna==3.3
mypy-protobuf==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
toolz==0.11.2
types-protobuf==3.19.12
urllib3==1.26.9
yapf==0.32.0
